--- versioned/tiered/tiered-impl/src/test/java/org/projectnessie/versioned/impl/condition/SetCommand.java
@@ -1,29 +1,0 @@
-/*
- * Copyright (C) 2020 Dremio
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.projectnessie.versioned.impl.condition;
-
-import org.immutables.value.Value.Immutable;
-import org.projectnessie.versioned.store.Entity;
-
-/**
- * Sample of a specific type of update command into which UpdateClauses are converted.
- */
-@Immutable
-abstract class SetCommand implements UpdateCommand {
-
-  abstract Entity getEntity();
-}
