--- versioned/tiered/tiered-impl/src/test/java/org/projectnessie/versioned/impl/condition/UpdateCommand.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (C) 2020 Dremio
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.projectnessie.versioned.impl.condition;
-
-/**
- * Sample interface or class into which UpdateClauses are converted.
- */
-interface UpdateCommand {
-  /**
-   * An enum encapsulating.
-   */
-  enum Operator {
-    // An operator to remove some part or all of an entity.
-    REMOVE,
-
-    // An operator to set some part or all of an entity.
-    SET
-  }
-
-  Operator getOperator();
-
-  ExpressionPath getPath();
-
-}
